+++ crypto/algo/algo.go
@@ -8,21 +8,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/crypto/keys/hd"
 )
 
-// PubKeyType defines an algorithm to derive key-pairs which can be used for cryptographic signing.
-type PubKeyType string
-
-const (
-	// MultiType implies that a pubkey is a multisignature
-	MultiType = PubKeyType("multi")
-	// Secp256k1Type uses the Bitcoin secp256k1 ECDSA parameters.
-	Secp256k1Type = PubKeyType("secp256k1")
-	// Ed25519Type represents the Ed25519Type signature system.
-	// It is currently not supported for end-user keys (wallets/ledgers).
-	Ed25519Type = PubKeyType("ed25519")
-	// Sr25519Type represents the Sr25519Type signature system.
-	Sr25519Type = PubKeyType("sr25519")
-)
-
 var (
 	// Secp256k1 uses the Bitcoin secp256k1 ECDSA parameters.
 	Secp256k1 = secp256k1Algo{}
